# mw-general-bff/apiAutogenerated/mw-general-bff/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  title: Masters way general-bff API
  version: "1.0"
servers:
- url: /general
paths:
...
  /wayTags:
    post:
      operationId: create-wayTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreateWayTagPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.WayTagResponse'
          description: OK
      summary: Create a new wayTag
      tags:
      - wayTag
      x-codegen-request-body-name: request
  /wayTags/{wayTagId}/{wayId}:
    delete:
      operationId: delete-wayTag
      parameters:
      - description: wayTag ID
        in: path
        name: wayTagId
        required: true
        schema:
          type: string
      - description: way ID
        in: path
        name: wayId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete wayTag by UUID
      tags:
      - wayTag
...
components:
  schemas:
...
    mw-general-bff_internal_schemas.CreateWayTagPayload:
      properties:
        name:
          type: string
        wayUuid:
          type: string
      required:
      - name
      - wayUuid
      type: object
...
    mw-general-bff_internal_schemas.WayTagResponse:
      example:
        name: name
        uuid: uuid
      properties:
        name:
          type: string
        uuid:
          type: string
      required:
      - name
      - uuid
      type: object
...